finally figured out communal app header
