+++ template/tsconfig.json
@@ -23,6 +23,12 @@
       "prisma/*": ["./prisma/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts",
+    "src/components/app/Header.tsx.old"
+  ],
   "exclude": ["node_modules"]
 }
